+++ extensions/smallrye-graphql/deployment/src/test/java/io/quarkus/smallrye/graphql/deployment/AbstractGraphQLTest.java
@@ -1,6 +1,6 @@
 package io.quarkus.smallrye.graphql.deployment;
 
-import static io.quarkus.jsonp.runtime.JsonProviderHolder.JSON_PROVIDER;
+import static io.quarkus.jsonp.JsonProviderHolder.JSON_PROVIDER;
 
 import java.io.IOException;
 import java.io.StringReader;
