+++ extensions/smallrye-graphql/runtime/src/main/java/io/quarkus/smallrye/graphql/runtime/SmallRyeGraphQLExecutionHandler.java
@@ -1,6 +1,6 @@
 package io.quarkus.smallrye.graphql.runtime;
 
-import static io.quarkus.jsonp.runtime.JsonProviderHolder.JSON_PROVIDER;
+import static io.quarkus.jsonp.JsonProviderHolder.JSON_PROVIDER;
 
 import java.io.IOException;
 import java.io.StringReader;
